Add ability to specify initial delay before different object types spawn.
