+++ Assets/Config/PitTrapConfig.asset
@@ -9,6 +9,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: abef545bf624e55d98a97fc712bd43fd, type: 3}
-  m_Name: SpawnEnergyTrigger
+  m_Script: {fileID: 11500000, guid: a3785d60bebf1bd33ab7fc71eb40cb3f, type: 3}
+  m_Name: PitTrapConfig
   m_EditorClassIdentifier: 
+  Enabled: 1
+  FrequencyMin: 4
+  FrequencyMax: 8
+  InitialDelay: 15
