+++ Assets/Config/PitTrapConfig.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 09e1458313d018ed18438acc09bc6216
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
