+++ Assets/Config/PitTrapConfig.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "PitTrapConfig", menuName = "ScriptableObjects/PitTrapConfig", order = 1)]
+public class PitTrapConfig : SpawnableConfig
+{
+
+}
