+++ Assets/Config/PitTrapConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a3785d60bebf1bd33ab7fc71eb40cb3f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
