+++ Assets/Config/SpawnableConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c5ba472f6f9ada46badea7915aeb25f7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
