+++ Assets/Config/StarConfig.asset
@@ -1,0 +1,22 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 40245e5900edce9708a8204de209b003, type: 3}
+  m_Name: StarConfig
+  m_EditorClassIdentifier: 
+  Enabled: 1
+  FrequencyMin: 6
+  FrequencyMax: 20
+  InitialDelay: 0
+  MinRotation: -720
+  MaxRotation: 720
+  MinSpeed: 2
+  MaxSpeed: 6
