+++ Assets/Config/StarConfig.asset.meta
