+++ Assets/Config/StarConfig.cs.meta
