+++ Assets/Racetrack/SpawnStarTrigger.asset
@@ -9,10 +9,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 40245e5900edce9708a8204de209b003, type: 3}
-  m_Name: PowerUpConfig
+  m_Script: {fileID: 11500000, guid: abef545bf624e55d98a97fc712bd43fd, type: 3}
+  m_Name: SpawnStarTrigger
   m_EditorClassIdentifier: 
-  MinRotation: -720
-  MaxRotation: 720
-  MinSpeed: 2
-  MaxSpeed: 6
