+++ Assets/Racetrack/SpawnStarTrigger.asset.meta
