+++ Assets/Racetrack/Star.prefab
@@ -16,7 +16,7 @@ GameObject:
   - component: {fileID: -3029167633070068567}
   - component: {fileID: 7625697157316688759}
   m_Layer: 0
-  m_Name: EnergyPowerUp
+  m_Name: Star
   m_TagString: Power Up
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
