+++ Assets/Racetrack/Star.prefab.meta
