+++ Assets/Scenes/RaceScene.unity
@@ -560,8 +560,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cd026295e510cb701b69f6c36d4eea49, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  frequencyMin: 6
-  frequencyMax: 20
+  config: {fileID: 11400000, guid: 34f3402c767d505aa8cb66a2ab3cddac, type: 2}
   trigger: {fileID: 11400000, guid: 0f54f645cb4d4f19dbf731f37297117e, type: 2}
 --- !u!114 &371023079
 MonoBehaviour:
@@ -603,8 +602,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cd026295e510cb701b69f6c36d4eea49, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  frequencyMin: 3
-  frequencyMax: 8
+  config: {fileID: 11400000, guid: 09e1458313d018ed18438acc09bc6216, type: 2}
   trigger: {fileID: 11400000, guid: daa59f33ec9fd55eaac52e3e06256007, type: 2}
 --- !u!1 &473285830
 GameObject:
